team_code/requirements.txt
google-auth-oauthlib==0.4.6
absl-py==1.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.12.5
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.4.1
cachetools==5.2.0
catalogue==1.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cycler==0.11.0
cymem==2.0.3
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dictor==0.1.10
dill==0.3.5.1
diskcache==5.4.0
docker-pycreds==0.4.0
einops==0.4.1
entrypoints==0.4
ephem==4.1.3
fastjsonschema==2.16.1
filelock==3.8.0
filterpy==1.4.5
fonttools==4.37.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
gym==0.17.2
h5py==3.7.0
huggingface-hub==0.9.1
hydra-core==1.3.2
idna==3.3
imageio==2.21.2
imgaug==0.4.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.10
jsonschema==4.15.0
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyter_packaging==0.12.3
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.15.1
kiwisolver==1.4.4
laspy==2.2.0
laszip==0.1.0
lazy-object-proxy==1.7.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
meson==0.55.3
mistune==2.0.4
munkres==1.1.4
murmurhash==1.0.2
nbclassic==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.6.3
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.1
oauthlib==3.2.0
omegaconf==2.2.3
open3d
opencv-python==4.6.0.66
opt-einsum==3.2.1
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
plac==1.1.3
platformdirs==2.5.2
preshed==3.0.2
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.19.4
psutil==5.9.1
py-trees==0.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pygame==2.1.2
Pygments==2.13.0
pylint==2.15.0
pyparsing==3.0.9
pyro-api==0.1.2
pyro-ppl==1.3.1
pyrsistent==0.18.1
python-dateutil==2.8.2
pytictoc==1.5.2
pytorch-lightning==0.7.1
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.1
rdp==0.8
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scikit-multiflow==0.5.0
scipy==1.7.3
Send2Trash==1.8.0
sentry-sdk==1.9.7
setproctitle==1.3.2
shap==0.35.0
Shapely==1.8.4
shortuuid==1.0.9
simple-watchdog-timer==0.1.1
six==1.14.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==2.3.0
srsly==1.0.2
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
thinc==7.4.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.7
tinycss2==1.1.1
tokenizers==0.12.1
tomli==2.0.1
tomlkit==0.11.4
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1+cu113
--extra-index-url https://download.pytorch.org/whl/cu113
torchaudio==0.12.1
torchmetrics==0.11.0
--extra-index-url https://download.pytorch.org/whl/cu113
torchvision==0.13.1+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
transformers==4.21.2
typed-ast==1.5.4
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
wandb==0.13.2
wasabi==0.7.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.0
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
zipp==3.8.1
py_trees==0.8.3